backend/pyproject.toml
[project]
name = "track-backend"
version = "0.1.0"
description = "Track backend API"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.116.0",
  "uvicorn[standard]>=0.35.0",
  "sqlmodel>=0.0.24",
  "email-validator>=2.2.0",
  "python-jose[cryptography]>=3.4.0",
  "passlib[bcrypt]>=1.7.4",
  "python-multipart>=0.0.20",
  "pydantic-settings>=2.10.1",
  "pyyaml>=6.0.2",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.4.0",
  "httpx>=0.28.1",
]

[tool.uv]
package = false
